Refactor tools
