+++ Context/Context/Tools/ToolDetailView.swift
@@ -4,9 +4,6 @@ import AppKit
 import ComposableArchitecture
 import ContextCore
 import Dependencies
-import GRDB
-import MarkdownUI
-import SharingGRDB
 import SwiftUI
 
 struct ToolDetailView: View {
@@ -29,12 +26,7 @@ struct ToolDetailView: View {
   @Dependency(\.mcpClientManager) private var mcpClientManager
   @Dependency(\.defaultDatabase) private var database
 
-  // View mode for response
-  enum ViewMode: String, CaseIterable {
-    case preview = "Preview"
-    case raw = "Raw"
-  }
-  @State private var viewMode: ViewMode = .preview
+  @State private var viewMode: ToolViewMode = .preview
 
   var body: some View {
     GeometryReader { geometry in
@@ -47,7 +39,8 @@ struct ToolDetailView: View {
       }
     }
     .sheet(isPresented: $showingFullDescription) {
-      fullDescriptionSheet
+      ToolHeaderView(tool: tool, showingFullDescription: $showingFullDescription)
+        .fullDescriptionSheet
     }
     .onAppear {
       loadFromToolState()
@@ -103,97 +96,8 @@ struct ToolDetailView: View {
   @ViewBuilder
   private var topPane: some View {
     VStack(spacing: 0) {
-      headerSection
+      ToolHeaderView(tool: tool, showingFullDescription: $showingFullDescription)
       argumentsSection
-    }
-  }
-
-  @ViewBuilder
-  private var headerSection: some View {
-    VStack(alignment: .leading, spacing: 16) {
-      HStack(alignment: .top, spacing: 12) {
-        toolInfo
-        Spacer()
-        annotationBadges
-      }
-    }
-    .padding(20)
-    .background(Color(NSColor.controlBackgroundColor))
-  }
-
-  @ViewBuilder
-  private var toolInfo: some View {
-    VStack(alignment: .leading, spacing: 4) {
-      Text(tool.name)
-        .font(.title2)
-        .fontWeight(.semibold)
-
-      if let description = tool.description {
-        VStack(alignment: .leading, spacing: 4) {
-          Text(description)
-            .font(.callout)
-            .foregroundColor(.secondary)
-            .textSelection(.enabled)
-            .lineLimit(3)
-
-          Button(action: {
-            showingFullDescription = true
-          }) {
-            Text("Show more")
-              .font(.caption)
-              .foregroundColor(.accentColor)
-          }
-          .buttonStyle(.plain)
-        }
-      } else {
-        Button(action: {
-          showingFullDescription = true
-        }) {
-          Text("Show more")
-            .font(.caption)
-            .foregroundColor(.accentColor)
-        }
-        .buttonStyle(.plain)
-      }
-    }
-  }
-
-  @ViewBuilder
-  private var annotationBadges: some View {
-    if let annotations = tool.annotations {
-      HStack(spacing: 8) {
-        if annotations.readOnlyHint == true {
-          AnnotationBadge(
-            label: "Read Only",
-            color: .green,
-            icon: "eye"
-          )
-        }
-
-        if annotations.destructiveHint == true {
-          AnnotationBadge(
-            label: "Destructive",
-            color: .red,
-            icon: "exclamationmark.triangle"
-          )
-        }
-
-        if annotations.idempotentHint == true {
-          AnnotationBadge(
-            label: "Idempotent",
-            color: .blue,
-            icon: "arrow.triangle.2.circlepath"
-          )
-        }
-
-        if annotations.openWorldHint == true {
-          AnnotationBadge(
-            label: "Open World",
-            color: .purple,
-            icon: "globe"
-          )
-        }
-      }
     }
   }
 
@@ -246,122 +150,16 @@ struct ToolDetailView: View {
 
   @ViewBuilder
   private var bottomPane: some View {
-    VStack(spacing: 0) {
-      responseHeader
-      Divider()
-      responseContent
-    }
-  }
-
-  @ViewBuilder
-  private var responseHeader: some View {
-    HStack(spacing: 12) {
-      Text("Response")
-        .font(.headline)
-
-      Spacer()
-
-      // Copy button when in Raw mode
-      if viewMode == .raw && shouldShowCopyButton {
-        CopyButton {
-          copyRawDataToClipboard()
-        }
-        .help("Copy raw JSON to clipboard")
-      }
-
-      runToolButton
-    }
-    .padding(.horizontal, 20)
-    .frame(height: 50)
-    .background(Color(NSColor.controlBackgroundColor))
-    .overlay(
-      // Centered toggle buttons
-      ToggleButton(selection: $viewMode)
+    ToolResponseView(
+      hasLoadedOnce: hasLoadedOnce,
+      toolResponse: toolResponse,
+      responseJSON: responseJSON,
+      responseError: responseError,
+      viewMode: $viewMode,
+      canCallTool: canCallTool,
+      isLoading: isLoading,
+      onRunTool: callTool
     )
-  }
-
-  @ViewBuilder
-  private var runToolButton: some View {
-    Button(action: {
-      callTool()
-    }) {
-      ZStack {
-        Image(systemName: "play.fill")
-          .font(.system(size: 16))
-          .foregroundColor(.accentColor)
-          .opacity(isLoading ? 0 : 1)
-
-        if isLoading {
-          ProgressView()
-            .controlSize(.small)
-            .scaleEffect(0.8)
-        }
-      }
-    }
-    .buttonStyle(.plain)
-    .disabled(!canCallTool || isLoading)
-    .help("Run Tool (⌘↩)")
-  }
-
-  @ViewBuilder
-  private var responseContent: some View {
-    if hasLoadedOnce {
-      if let response = toolResponse {
-        Group {
-          switch viewMode {
-          case .preview:
-            let messages = response.content.map { content in
-              ToolResponseMessage(content: content)
-            }
-            MessageThreadView(messages: messages)
-          case .raw:
-            ToolRawDataView(
-              responseJSON: responseJSON,
-              responseError: responseError
-            )
-          }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-      } else if responseError != nil {
-        // Error state
-        Group {
-          switch viewMode {
-          case .preview:
-            if let error = responseError {
-              JSONRPCErrorView(error: error)
-            }
-          case .raw:
-            ToolRawDataView(
-              responseJSON: responseJSON,
-              responseError: responseError
-            )
-          }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-      } else {
-        ContentUnavailableView(
-          "No Response",
-          systemImage: "function",
-          description: Text("Tool call completed but no response was returned")
-        )
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-      }
-    } else {
-      ContentUnavailableView(
-        "No Response",
-        systemImage: "function",
-        description: Text("Enter arguments and click the ▶ button to call the tool")
-      )
-      .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-  }
-
-
-  private func calculateIdealHeight() -> CGFloat {
-    let baseHeight: CGFloat = 160  // Header + padding
-    let propertiesCount = tool.inputSchema.properties?.count ?? 0
-    let propertyHeight: CGFloat = 50  // Height per property
-    return baseHeight + (CGFloat(min(propertiesCount, 5)) * propertyHeight)
   }
 
   private var canCallTool: Bool {
@@ -385,18 +183,6 @@ struct ToolDetailView: View {
     }
 
     return validationErrors.isEmpty
-  }
-  
-  private var shouldShowCopyButton: Bool {
-    // Show copy button if we have raw JSON or if there's an error
-    responseJSON != nil || responseError != nil
-  }
-  
-  private func copyRawDataToClipboard() {
-    RawDataView.copyRawDataToClipboard(
-      responseJSON: responseJSON,
-      responseError: responseError
-    )
   }
 
   private func initializeParameterValues() {
@@ -550,11 +336,11 @@ struct ToolDetailView: View {
       } catch {
         // Store the actual error object
         responseError = error
-        
+
         // Clear the tool response - we'll show error view instead
         toolResponse = nil
         responseJSON = nil
-        
+
         hasLoadedOnce = true
         isLoading = false
 
@@ -571,156 +357,4 @@ struct ToolDetailView: View {
     }
   }
 
-  @ViewBuilder
-  private var fullDescriptionSheet: some View {
-    VStack(spacing: 20) {
-      HStack {
-        Text(tool.name)
-          .font(.title2)
-          .fontWeight(.semibold)
-
-        Spacer()
-
-        Button("Done") {
-          showingFullDescription = false
-        }
-        .keyboardShortcut(.defaultAction)
-      }
-
-      ScrollView {
-        VStack(alignment: .leading, spacing: 16) {
-          if let description = tool.description {
-            Markdown(description)
-              .markdownTextStyle {
-                ForegroundColor(.primary)
-              }
-              .textSelection(.enabled)
-              .frame(maxWidth: .infinity, alignment: .leading)
-          }
-
-          if let properties = tool.inputSchema.properties, !properties.isEmpty {
-            Divider()
-
-            VStack(alignment: .leading, spacing: 8) {
-              Text("Arguments")
-                .font(.headline)
-
-              let requiredSet = Set(tool.inputSchema.required ?? [])
-              let sortedProperties = properties.sorted { $0.key < $1.key }
-
-              ForEach(sortedProperties, id: \.key) { key, schema in
-                VStack(alignment: .leading, spacing: 4) {
-                  HStack {
-                    Text(key)
-                      .font(.subheadline)
-                      .fontWeight(.medium)
-
-                    if requiredSet.contains(key) {
-                      Text("Required")
-                        .font(.caption2)
-                        .fontWeight(.medium)
-                        .padding(.horizontal, 6)
-                        .padding(.vertical, 2)
-                        .background(
-                          RoundedRectangle(cornerRadius: 4)
-                            .fill(Color.red.opacity(0.2))
-                        )
-                        .foregroundColor(.red)
-                    }
-
-                    if let type = extractSchemaType(from: schema) {
-                      Text(type)
-                        .font(.caption2)
-                        .fontWeight(.medium)
-                        .padding(.horizontal, 6)
-                        .padding(.vertical, 2)
-                        .background(
-                          RoundedRectangle(cornerRadius: 4)
-                            .fill(Color.blue.opacity(0.2))
-                        )
-                        .foregroundColor(.blue)
-                    }
-                  }
-
-                  if let desc = extractDescription(from: schema) {
-                    Markdown(desc)
-                      .markdownTextStyle {
-                        ForegroundColor(.secondary)
-                      }
-                      .font(.caption)
-                      .textSelection(.enabled)
-                  }
-                }
-                .padding(.vertical, 4)
-              }
-            }
-          }
-
-          if let annotations = tool.annotations {
-            Divider()
-
-            VStack(alignment: .leading, spacing: 8) {
-              Text("Annotations")
-                .font(.headline)
-
-              HStack(spacing: 12) {
-                if annotations.readOnlyHint == true {
-                  AnnotationBadge(
-                    label: "Read Only",
-                    color: .green,
-                    icon: "eye"
-                  )
-                }
-
-                if annotations.destructiveHint == true {
-                  AnnotationBadge(
-                    label: "Destructive",
-                    color: .red,
-                    icon: "exclamationmark.triangle"
-                  )
-                }
-
-                if annotations.idempotentHint == true {
-                  AnnotationBadge(
-                    label: "Idempotent",
-                    color: .blue,
-                    icon: "arrow.triangle.2.circlepath"
-                  )
-                }
-
-                if annotations.openWorldHint == true {
-                  AnnotationBadge(
-                    label: "Open World",
-                    color: .purple,
-                    icon: "globe"
-                  )
-                }
-              }
-            }
-          }
-        }
-        .padding(.vertical)
-      }
-    }
-    .padding(20)
-    .frame(width: 600, height: 400)
-  }
-
-  private func extractSchemaType(from schema: JSONValue) -> String? {
-    if case .object(let obj) = schema,
-      case .string(let typeStr) = obj["type"]
-    {
-      return typeStr
-    }
-    return nil
-  }
-
-  private func extractDescription(from schema: JSONValue) -> String? {
-    if case .object(let obj) = schema,
-      case .string(let desc) = obj["description"]
-    {
-      return desc
-    }
-    return nil
-  }
 }
