+++ Context/Context/Tools/ToolViewMode.swift
@@ -1,0 +1,10 @@
+// Copyright © 2025 Indragie Karunaratne. All rights reserved.
+
+import Foundation
+
+// View mode for tool response
+enum ToolViewMode: String, CaseIterable {
+  case preview = "Preview"
+  case raw = "Raw"
+}
+
